+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   InlineList: add `wrap` prop to activate line wrap on overflow
+-   Thumbnail: add `objectFit` prop to control how the image fit in a constrained aspect ratio.
+    Defaults to `cover` to scale & crop the image (like before).
+    Can be changed to `contain` to avoid cropping the image (aka letterboxing).
 
 ## [3.7.0][] - 2024-04-29